+++ src/connectors/lists/list.card.js
@@ -32,6 +32,10 @@ export const ListCard = ({ id }) => {
     dispatch(sendSnowplowEvent('shareable-list.public-link.copy.all-lists', analyticsData))
   }
 
+  const onOpenPublicUrl = () => {
+    dispatch(sendSnowplowEvent('shareable-list.public-link.open.all-lists', analyticsData))
+  }
+
   const onOpenList = () => {
     dispatch(sendSnowplowEvent('shareable-list.open', analyticsData))
   }
@@ -51,6 +55,7 @@ export const ListCard = ({ id }) => {
         openUrl={`/lists/${externalId}`}
         onItemInView={onItemInView}
         onCopyPublicUrl={onCopyPublicUrl}
+        onOpenPublicUrl={onOpenPublicUrl}
         onOpen={onOpenList}
         isInternalItem={true}
         listStatusInfo={listStatusInfo}